Force default Measurement to not use request settings
When default Measurement is read from file, ensure that it does not use
request settings. Otherwise calling Measurement.to_file would only write
.info file (a change introduced in fc68f07)

This is a backwards compatibility fix for requests that were made with an
older version of Potku. Value of use_request_settings flag used to be written
to .profile, but now its written to .info. When Potku cannot find the value,
it defaults to True.