Update value when text is changed
The value of the ScientificSpinBox is now updated whenever user
enters some valid text into the element. This feature was broken
when the class was refactored to subclass QDoubleSpinBox instead
of QAbstractSpinBox.

There may be better ways to update the value, for example
using editingFinished signal or overriding valueFromText method but
I was not able to get these working.