Fix ScientificSpinBox sizing issue on Windows (#214)
ScientificSpinBoxes appeared smaller than other spin boxes on
Windows. This was fixed by setting the value and ranges of the
ScientificSpinBox before setting the validator of the LineEdit.

Validator calls LineEdit.setStyleSheet to change the background
color to white, yellow or red depending on the validity of the input
value. Calling LineEdit.setStyleSheet before the value is set seems
to break something, at least on Windows. Simply changing the call
order fixes this.